+++ app/javascript/dashboard/routes/dashboard/settings/inbox/Settings.vue
@@ -241,6 +241,13 @@
           >
             <woot-code :script="inbox.web_widget_script"></woot-code>
           </settings-section>
+
+          <settings-section
+            :title="$t('INBOX_MGMT.SETTINGS_POPUP.HMAC_VERIFICATION')"
+            :sub-title="$t('INBOX_MGMT.SETTINGS_POPUP.HMAC_DESCRIPTION')"
+          >
+            <woot-code :script="inbox.hmac_token"></woot-code>
+          </settings-section>
         </div>
       </div>
     </div>
