+++ app/javascript/widget/store/modules/contacts.js
@@ -2,16 +2,23 @@ import ContactsAPI from '../../api/contacts';
 import { refreshActionCableConnector } from '../../helpers/actionCable';
 
 export const actions = {
-  update: async (_, { identifier, user: userObject }) => {
+  update: async ({ dispatch }, { identifier, user: userObject }) => {
     try {
       const user = {
         email: userObject.email,
         name: userObject.name,
         avatar_url: userObject.avatar_url,
+        identifier_hash: userObject.identifier_hash,
       };
       const {
         data: { pubsub_token: pubsubToken },
       } = await ContactsAPI.update(identifier, user);
+
+      if (userObject.identifier_hash) {
+        dispatch('conversation/clearConversations', {}, { root: true });
+        dispatch('conversation/fetchOldConversations', {}, { root: true });
+      }
+
       refreshActionCableConnector(pubsubToken);
     } catch (error) {
       // Ingore error
