+++ app/views/widget_tests/index.html.erb
@@ -1,5 +1,15 @@
 <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=0" />
 
+
+<%
+  user_id = 1
+  user_hash = OpenSSL::HMAC.hexdigest(
+    'sha256',
+    @web_widget.hmac_token,
+    user_id.to_s
+  )
+
+%>
 <script>
 
 window.chatwootSettings = {
@@ -24,6 +34,11 @@ window.chatwootSettings = {
 })(document,"script");
 
 window.addEventListener('chatwoot:ready', function() {
-  console.log(window.$chatwoot)
+  console.log(window.$chatwoot);
+  window.$chatwoot.setUser('<%= user_id %>', {
+    identifier_hash: '<%= user_hash %>',
+    email: 'jane@acme.inc',
+    name: 'Jane Doe'
+  });
 })
 </script>
